adjust file configuration using blueprint-ok
