+++ templates/index.html
@@ -4,7 +4,7 @@
     <meta http-equiv="X-UA-Compatible" content="IE=edge">
     <meta name="viewport" content="width=device-width, initial-scale=1.0, height=device-height">
     <link rel="shortcut icon" href="#"/>      <!-- favicon.ico問題 -->
-    <link rel="stylesheet" href="/css/style.css">
+    <link rel="stylesheet" href="/css/index.css">
     <title>Taipei Trip 台北一日遊</title>
 </head>
 <body>
